fix: resolve all Kotlin compilation errors (22 files)
Fix compilation errors reported in build log:

1. ExpressionParser.kt: convert eval() from expression body to block body
   so 'return' inside when-branches is legal.

2. Tune.kt: replace map { (k, v) -> k to v ... }.toMutableMap() with
   mapValues { ... }.toMutableMap() for proper type inference.

3. TuneSerializer.kt: switch to kotlinx.serialization.encodeToString
   (reified extension) from the unresolved .json.encodeToString.

4. DialogRenderer.kt: add import androidx.compose.foundation.text.KeyboardOptions
   so OutlinedTextField's keyboardOptions param resolves.

5. LibreTuneApp.kt:
   - Add imports findStartDestination, NavigationDrawerItem, HorizontalDivider,
     fillMaxWidth, size.
   - Add new composable MenuTreeDrawer to replace the missing MenuTree call.
   - Fix TableEditorScreen parameter name -> tableName.
   - Use block-style popUpTo { saveState = true } instead of trailing-lambda form.
   - Use collectAsState() pattern for menuTree + hasDefinition.

6. MenuTreeViewModel.kt:
   - Annotate with @HiltViewModel + @Inject constructor so hiltViewModel()
     can resolve it at runtime.
   - Change constructor param EcuDefinition? to EcuDefinition (non-null injection).
   - Update init guard accordingly.

7. VeAnalysisScreen.kt:
   - Rename nested it lambda params to newValue to avoid shadowing.
   - Use androidx.compose.ui.graphics.lerp(start, stop, fraction) top-level fn
     instead of Color.lerp (which does not exist).
   - Fix smart cast on state.results via local val currentResults.

8. VeAnalysisViewModel.kt:
   - .tables.firstOrNull -> .tables.values.firstOrNull (Map requires .values).
   - .curves.firstOrNull -> .curves.values.firstOrNull.
   - resolveChannel(...) ?: "" for non-null String field rpm.

9. AutoTuneScreen.kt:
   - Change SettingsPanel.onUpdate signature to
     ((AutoTuneSettings) -> AutoTuneSettings) -> Unit so
     viewModel::updateSettings matches.
   - Rename nested it lambda params.
   - Replace illegal nested items { item { ... } items { ... } } inside
     LazyVerticalGrid with a single flat items(rows * (cols + 1)) block using
     index-based row/col derivation.

10. SensorCalibrationViewModel.kt:
    - Replace writer.setPosition(offset); writer.writeValue(...) with
      writer.writeValueAt(offset, ...) since setPosition does not exist on
      ByteOrderWriter.

11. CurveEditorScreen.kt:
    - textMeasurer.measure(text, fontSize = 10.sp) ->
      textMeasurer.measure(text, style = TextStyle(fontSize = 10.sp)).
    - measured.size.height / 2 -> / 2f (avoid Int/Float ambiguity).

12. CurveEditorViewModel.kt:
    - curve.page -> curve.binsPage (field does not exist on CurveDefinition).
    - Drop curve.xAxis reference (decode bins directly from binsOffset).
    - curve.xOutputChannel -> null (not a field of CurveDefinition).
    - Cast (position.x - chartBounds.left).toDouble() / chartBounds.width.toDouble()
      so coerceIn(0.0, 1.0) works.

13. DashboardViewModel.kt: add SYNCING branch to make when exhaustive.

14. BarGauge.kt: coerce Double ratio to Float (.toFloat().coerceIn(0f, 1f)).

15. DatalogChartScreen.kt:
    - Move imports awaitEachGesture/awaitFirstDown to
      androidx.compose.foundation.gestures.*.
    - Add imports drawLayer, Stroke, clipRect.
    - Replace it.pressed with it.isPressed (Compose 1.7 API).
    - Fix smart cast on state.crosshairTime via state.crosshairTime!!.
    - Cast size.width to Float where Float args are required.

16. EcuConsoleViewModel.kt: ArrayDeque<String>(capacity = 64) ->
    ArrayDeque<String>(64) (positional arg).

17. ToothLoggerScreen.kt:
    - Add import com.ztune.libretune.core.ecu.ToothLoggerTransport.
    - Pass TextMeasurer to drawCenteredPlaceholder/drawAxisLabels instead of
      calling TextMeasurer() ctor (no public no-arg ctor).
    - Use rememberTextMeasurer() in ToothTimingDiagram composable.
    - measure() calls use TextStyle(fontSize = 10.sp).

18. FrontPageScreen.kt:
    - Remove subtitle = { ... } param (not in Material3 TopAppBar API),
      move sig Text into title = { Column { ... } }.
    - Rename val _ = channelStore.channels.collectAsState() to
      val channelState = ... (_ is reserved in Kotlin).

19. TuneHistoryScreen.kt: HistoryRestore (unresolved) -> Restore icon.

20. HelpViewerScreen.kt:
    - Remove bulletListType = BulletListType() (not a Text param).
    - Move textIndent into style = MaterialTheme.typography.bodyMedium.copy(...).
    - Use 16.sp (TextUnit) instead of 16.dp for restLine.

21. TableColorUtils.kt: cast t / 0.5 to Float for Color(red: Float, ...).

22. TableEditorViewModel.kt:
    - tbl.xOutputChannel / tbl.yOutputChannel -> null (not TableDefinition fields).
    - Fix TableOperations.scaleCells(cells, factor, mutable) arg order.
    - Fix TableOperations.interpolateCells(table, selectedCells, values, rows, cols)
      signature (was missing values param).
    - Fix TableOperations.setCellsEqual(cells, emptyList(), mutable) (was passing
      mutable twice).
    - Fix TableOperations.addOffset(cells, offset, mutable) arg order.
    - Change getSelectedCells() return type to List<Pair<Int, Int>>? (cell coords
      only, not values).

+++ app/src/main/java/com/ztune/libretune/ui/screens/dashboard/gauges/BarGauge.kt
@@ -109,7 +109,7 @@ private fun HorizontalBarContent(
     onSurfaceVariant: Color
 ) {
     val range = config.max - config.min
-    val fraction = if (range != 0.0) ((value - config.min) / range).coerceIn(0.0, 1.0) else 0f
+    val fraction = if (range != 0.0) ((value - config.min) / range).toFloat().coerceIn(0f, 1f) else 0f
     val animatedFraction by animateFloatAsState(
         targetValue = fraction,
         animationSpec = androidx.compose.animation.core.tween(durationMillis = 150)
@@ -218,7 +218,7 @@ private fun VerticalBarContent(
     onSurfaceVariant: Color
 ) {
     val range = config.max - config.min
-    val fraction = if (range != 0.0) ((value - config.min) / range).coerceIn(0.0, 1.0) else 0f
+    val fraction = if (range != 0.0) ((value - config.min) / range).toFloat().coerceIn(0f, 1f) else 0f
     val animatedFraction by animateFloatAsState(
         targetValue = fraction,
         animationSpec = androidx.compose.animation.core.tween(durationMillis = 150)
